Refactoring after change of module structure

+++ modules/user/user-usecase/src/main/java/com/kosinski/user/usecase/impl/AccountNumberNotFound.java
@@ -1,7 +1,7 @@
 package com.kosinski.user.usecase.impl;
 
 class AccountNumberNotFound extends RuntimeException {
-    AccountNumberNotFound(int accNumber) {
+    AccountNumberNotFound(long accNumber) {
         super(String.format("Account number %s doesn't exist!", accNumber));
     }
 }
